Commit of many improvements before reverting to previous commit to debug hypertarget location issue
This commit includes numerous improvements, particularly to
figure/table/listings environments and their entries in the table of
contents and their tcb styles and appearance.  However, hyperlinks no
longer point to the hypertarget anchor text like they did but to the top
of the page instead.  The glossaries navigation panel also do not point
to the correct position either.

I can't seem to find what modification caused this change in hypertarget
destination and attempts to copy the contents of the previous commit's
files did not solve this, so now I am committing the current status of
the files with all of the improvements and I will now revert all files
back to their versions from the previous commit in the hopes that the
links work properly in that commit and I can either find the cause of
the issue or at worst I can simply add the important change from the
current version of the files to the reverted files