Restored and further improved packages after fixing hyperlink issue
The issue with the hyperlinks pointing to the wrong location was due to
the SumatraPDF settings.  The links were not working as expected in
Adobe Acrobat either so there may be additional settings/configurations
changes needed, but it does not appear to be an issue caused by the
LaTeX implementations.

After determining the cause of the issue, the files from the previous
commit were restored from those from the commit before it, then
additional improvements and simplifications were made as part of the
continuing development of the TOC entries and tcolorbox titles and the
hyperlinks connecting these to each other