More tcb standardization, code listing improvements, dual acronym/glossary entries, and auto math font in text contexts
(1)  Continued standardizing tcolorbox styles, adding
plain/single/double lined border style options which automatically
format frame/title borders by including a 'tcbborders' call in a style
definition and passing the desired border width (boxrule), border line
type (plain/single/double), and color.  The environments that currently
incorporate this into their style definitions are IBPStyle, MatlabStyle,
tcbbashboxStyle, and tcbFigboxStyle.
(2) Continued improving the code listing environments for Matlab and
terminal (bash) listings.  The terminal listings were modified to
conform to the new tcbborder style option while maintaining the required
formatting for the listing comments (segmentation/tcblower regions).
The Matlab environment was also modified for use of tcbborder, but this
included modifications to the caption style/format.  These also required
modifications to the handling of their counter variables, particularly
for the Matlab caption which displays an incremented counter value by
default, requiring appropriate counter values to be saved/set manually
to maintain proper numbering in the title/caption/toc.  In both cases
the 'number within' tcb option was used to maintain a total count of all
code listing types for the toc numbering while maintaining a count of
each individual type.

This also factored in to the modifications of the tcb title and toc list
entry field modifications to ensure proper numbering and text entries in
these fields and additional code/formatting separate from these fields
so the title/list entry fields can be entered with the same set of
commands/macros/text, providing the basis for standardization of these
field entries and those of other environments (tcbfigbox, tcbfigure,
tcbfigboxfigure, tcbIBP, etc.) so commands/macros/defs can be used as
entries to these fields to automatically format them as desired with a
single consistent scheme.  These will be designed and used in the same
way as the other existing commands/macros which automatically format tcb
titles and toc entries and create two way links to/from the tcb titles
and toc entries as has been done for tcbfigbox environments.  These
standardized macros/commands are created automatically with the help of
the tocloft package by calling its toc list creation command, which
itself defines a number of commands automatically, and adding additional
command creation macros to this so a new toc entry list type can be
created along with a number of commands/macros for this list type whose
names are of the form cftX..., where X is the new list type name, based
on only a small # of arguments.  The currently existing standardized
entry macros/commands currently work with standard environments as
tcolorbox options and will wither be modified to also work as list
entry/title field arguments of tcolorbox specific environment
definitions or additional macros/commands will be created based off
their design which can be used in these contexts.
(3) The dual glossary/acronym entry definition was modified so in
addition to the "see ..." hyperlink from the "List of Acronyms" to the
"Glossary", now there is also a similar link in the opposite direction.
(4) When math is added to a text region context, the text
formatting/style commands (\bfseries etc.) do not alter the
formatting/style of the proceeding math text/fonts, resulting in normal
math font formatting/style inside bold/italic text contexts.  To
automatically format math fonts when inside a bold text context, a new
"\bfmseries" command was created so the math is also bold in such cases.
Additional commands will also be created for other text font styles as
well.  This is especially noticeable in the case of tcb title and toc
entries, where text is typically bold and math text which is not bold
too stands out, but this has now been resolved.