Added package 'X' patching schemes via 'my-patches.sty' or explicit 'X-patched.sty' packages
(1) Created 'my-patches' subpackage to load/apply a patch 'my-X-patch'
of a package 'X' for each of the patches provided in 'my-latex'.  These
modify and potentially add new macros/command/environments and also
provide an interface by which the behavior/appearance of content
generated by the package can be centrally controlled through a set of
register variables and control sequences specific to the (sub)package.
Although these can be centrally controlled explicitly, an additional set
of register variables and control sequences are also created for content
of a particular type/category (e.g. 'tcolorbox'/'tikz' environments,
front/back matter, etc.) and these are coupled to those that interface
with the (sub)packages generating content of this type/category, thereby
allowing users to modify the behavior/appearance of this content through
the type/category specific register variables and control sequences
instead of having to modify those for each of the corresponding
(sub)packages individually.
(2) The adapted versions of packages like 'glossaries', 'nomencl', and
'imakeidx' are also being redesigned to conform to this "package +
patch" scheme so people can use 'my-latex' and incorporate it into
existing documents without causing errors/clashes with usage of the
underlying base packages and no additional restrictions are created
regarding the order in which packages are loaded.  In addition, the
resulting patches of these widely/commonly used packages will include an
option to safely exclude the patch through additional 'my-latex' package
options, which are then forwarded to the corresponding (sub)package to
prevent application of the patch, thereby allowing the unmodified base
package to be used if desired.  Note that since the patches provide the
interface by which the behavior/appearance is centrally controlled, the
content generated using the base package will no longer be controllable
with this scheme and will not match the content generated by other
(sub)packages of its type/category.  This redesign and separation into
"package + patch" implementation has only been completed for 'imakeidx'
thus far but this will be done for the others in the near future.  To
make it possible to control the package options of such packages without
requiring users to modify 'my-latex', corresponding package options are
added to 'my-latex' and 'my-patches' and their values are then forwarded
to the package itself through a hierarchy of \PassOptionsToPackage calls
or corresponding Booleans and/or control sequences
(3) In cases where a patch must be applied immediately after loading the
corresponding base package, the loading/patching of the package is
encapsulated in a separate package 'X-patched' to ensure these 2 steps
happen sequentially by simply loading this package instead of the
underlying base package.