Replaced \ProcessOptions w/ \ProcessKeyvalOptions and corresponding key definitions
To provide support for key=value package options, the tools provided by
the 'kvoptions' package were incorporated into 'my-latex' and used to
replace the usage of \DeclareOption and \ProcessOptions, with the
corresponding \Declare<String/Bool/Void/etc.>Options and
\ProcessKeyvalOptions.  This involved usage of \SetupKeyvalOptions to
change the key control sequence 'prefix' to 'mylatex' since the '-' in
'my-latex' used as the prefix by default is not permitted in control
sequence names, e.g. the Boolean switch statements \if<prefix>@<key>.
The package options created to provide access to the 'imakeidx' package
options via 'my-latex' are defined using \DeclareVoidOption so the
\DeclareOption syntax can be used with \PassOptionsToPackage to set the
corresponding 'imakeidx' package options accordingly.