Added key/value package option subpackage
The 'kvoptions' package provides the ability to define key/value package options by building on the 'keyval' framework that provides key/value options in a macro/command/environment context.  Existing package option definitions/management scheme provided by the LaTeX kernel are also replaced with the equivalent 'kvoptions' implementations.  However, 'kvoptions' does not provide an equivalent of the  'keyval' package's '\define@choicekey' command and, hence, there is no mechanism to ensure an acceptable value is given to a string key.  Therefore, the 'pkgopts' subpackage added in this commit provides a '\DeclareChoiceOption' command that generates a key using the '\DeclareStringOption' command and restrict its value to a user defined set of acceptable "choices".  These acceptable "choices" are passed as a comma separated list and after processing the given package options, a "choice: key processing command also provided in this commit is executed to determine if they key was used and verify its value is a member of the acceptable value list, setting the key's corresponding usage Boolean true if it was used and throwing a package error if an invalid value was given.  The "Choice" key processing command must currently be executed for each such key individually, with the key name passed as an argument, but the implementation will be modified to maintain a list of all defined  "Choice" keys so these can all be processed automatically with a single command.  This will likely be combined with '\ProcessKeyvalOptions' to create a single command that performs all key processing.

In prevous versions of the 'my-latex' package implementation the package options were defined/processed using tools provided by the LaTeX kernel but the implementation of this supplemental key/value package option management scheme requires 'kvoptions' and several other packages, hence, these must be loaded prior to any \documentclass definitions or any subpackages are loaded.  This is a departure from the original structure of the package where the package options are used to configure the \documentclass before loading the subpackages, including the subpackage which loads all external/3rd party packages are loaded.  To support the implementation of the "Choice" key definition/management scheme and prevent duplicate package loads, a set of package loading commands are provided to load these required packages, maintain a list of currently loaded packages, and replace \RequiredPackage calls throughout the package with a new package loading command that only loads a package if it has not yet been loaded.  These are primarily implemented using base LaTeX kernel commands to minimize the number of required packages, but a list of these and those required for the the key/value definition/processing commands are predefined as comma separated lists and passed to the package loading commands.  The package loading commands can be used to load these package lists, skipping any that have already been loaded, and can also be used to load individual packages, thereby providing a single command to perform all package loading.