Encapsulated package option processing and documentclass configuration
Modified pkgopts to maintain a list of defined "Choice" options and
process all of these with a single command instead of as indivudal
commands.  This also combines this choice option processing command and
the \ProcessKeyvalOptions command into a single command to perform all
package option processing with a single command.

This commit also adds a \SetDocumentClass command to respond to package
options delated to the documen class, allowing users to use any document
class but giving preference to the Boolean package options for
report/article class.  In addition, the package option info for all
choice keys can now also be printed with a single command.