Simplified 'pkgopts', created single 'newkeyenvironment' document environment, and expanded 'frontbackmatter' '\cftX...' command
Redesigned 'pkgopts', which was already designed initially using as many
standard LaTeX kernel register variables and control sequences  (e.g.
\newcommand, \def, etc.) to minimize the dependence on external
packages.  However, these definitions still required 'forarray',
'xstring', and 'xparse' in some cases, but the corresponding
'\ForEachX', '\IfEq', 'IfStrEq', etc. commands have now been replaced
with '\@for' and '\ifx' macros, respectively, and the package loading
commands are now completely written with standard LateX kernel macros
and the remaining commands associated with the "Choice" key
definitions/management now only require the 'kvoptions' package they
build upon and 'etoolbox' in their implementation.  The 'kvoptions'
package cannot be avoided as these commands are meant to extend the
functionality of this package and it is used directly for most of the
other package options.  On the other hand, the '\ifstrempty' and
'\ifdefstrempty' commands used in their implementation can be replaced
with '\ifx' and similar commands in the future, though this is not a
high priority and if any difficulties arise with this, it may be delayed
possibly even permanently.

Also implemented in this commit is a redesign of the foundation of the
'documentation' and 'homework' environments containing the standard
document' environment and extending its functionality with an optional
'coverpage', left/right header, and PDF document info/properties.
Although these 2 environments have not been replaced yet, a
'newkeyenvironment' named 'thedocument' has been created which will
eventually be used to provide the functionality of both of these
environments with a single environment using a set of corresponding
option keys.  The 'enum' key type will be used to select which of the
document environments to use (including the standard 'document') and
additional keys will be used to provide title/subtitle/author (for
'documentation') as well as course #/name/instructor/semester and
assignment # + Chapter/problem #s (for 'homework').

The '\cftX...' command expansion was focused on replacing the explicit
title, list entry, title blend, and tcb/toc title font commands of the
various command/environments whose usage generate entries in the Table
of Contents.  As seen in previous commits, these are designed to expand
upon the 'tocloft' package with a set of generic functions \cftX...
which create specific commands for particular values of 'X' and these
are then encapsulated in a '\cftXcreate' command that calls each of
these generic functions, resulting in construction of each of these
commands for each list entry type.  This has now been redesigned such
that the bulk of the supplemental commands and settings/configurations
of the newly created properties and existing 'tocloft' properties are
performed by the newly created \cftXconfigure command.  These have then
been used to replace the majority of the explicit ToC entries and
'tcolorbox' properties, though there are a few which still contain
explict usage but these correspond to command/environments which aren't
currently used in any documents.

In the process of these changes/additions, many of the universal/general
purpose control sequence definitions were moved to 'pkgopts' and several
additional control sequences were added for argument parsing and control
sequence info queries.  Although these have many applications, the
primary use was to be able to strip the '\' from a control sequence and
extract its name as a string.