Expanded title/toc 2way hyperlinks and added graphics tools
The \cftX... commands configuring and generating the commands for each
application X were expanded, improved, and simplified for more universal
implementation and usage and more contexts/applications now use these to
create 2way links between the TOC and environment titles.  This also
included an inproved design of the shadow text commands which now detect
math environment content and apply shadows to these as well.  This
involved redesign and addition of font control sequences in my-fonts.
One flaw remains: math content cannot come last in such shadow text
contexts, hopefully this can be resolved in the future.

Some graphics tools were also added to create 2D cubes with colored
faces, which may vary on each face, and an assembly command which allows
numerous cubes to be stacked in rows/columns of various heights.  These
can then be rotated and their colors/sizes changed, thereby allowing
these to be rectangular instead.  The coloring of the faces must be
performed in a particular order and for some x/y/z rotations, these do
not display as desired.  There is room for improvement here and other
rotational tools such as rotation matrices can be employed, but with
appropriate choice of row/column/height and rotations, the desired
structure should be possible to create.