Significantly enhanced var # arg parsing, conditionals, num->text, and doc sectioning
The previous macros to parse csv argument lists of the form <...>=<...>
have now been reduced to a small number of flexible macros that accept a
variable number of arguments in each group, with each argument
parsed/stored in /argx/argX where x/X are its corresponding indices in
roman/Roman.  The supported operations on these pair groups are
/setlength, /def, and /let.  Commands and code segments can also be
passed as arguments along with a comma/space separated or braced group
of arguments with the command taking 1+ of the arguments in the argument
group and/or code segments which use 1+ of these arguments.  This allows
arguments to be passed to a macro/environment as a group instead of
individually to reduce the # of command arguments to avoid the limit of
9 arguments while still allowing macros/code to be passed to the same
command which uses individual arguments but requires its dependent
arguments to be parsed/separated from the group before they can access
them.  These enhancements make use of the \getargsC macro from the
'readarray' package as well as \getcsvargs and \getvargs macros defined
as adaptations of \getargsC to support csv and braced argument groups.

Significant enhancements to the manipulation, setting, reading/printing,
and execution dependent on conditional variables of various types: TF,
True/False, binary (0/1), counter (0/1), if, toggle, and xparse
Booleans.  There are control sequences to set, print, compare, convert
their type, store TF values, and a number of macros which use these to
execute code based on their value, including implementations of
AND/OR/XOR and other conditionals that depend on multiple conditional
values of potentially mixed type.

The macros that convert numbers to their text names/ordinals have also
been improved and reduced to a small number of shared macros such that
they can now be used to name and define new control sequences since the
various packages that support such number to text conversion do not
support the usage of the resulting text to name control sequences.

The document sectioning commands have also been modified once again.
The vertical spacing before/after individual and consecutive document
sections has been a significant challenge and was previously resolved,
but it became apparent that text inserted before/between/after these was
not receiving the appropriate vertical spacing since their presence
outside of control sequences requires the document sectioning commands
to insert the appropriate spacing without direct
interaction/communication with the other document sections.  This now
appears to be resolved for the document sectioning types and directly
inserted text but there will need to be additional modifications to some
of the other macros/environments that were previously designed for
proper spacing assuming the previous design of the document sectioning
commands (e.g. problem/solution, legend, figure/table/code/definition
boxes, text/equation list boxes, etc.).

Lastly, the numerous control sequences added to extend the 'tocloft'
package and simplify/standardize the usage and behavior/appearance of
front/back matter content (Table of Contents, Glossary, Index, List of
Symbols/Abbreviations/Acronyms, List of
Problems/Figures/Tables/Scripts/Functions, etc.) have been separated
from the 'mainmatter' package and placed in the new package
'tocloft-patched', primarily to reduce the clutter that had accumulated
in 'mainmatter' due to the sheer volume of control sequences created for
this purpose.