Code separated further, doc/toc/coverpage settings encapsulated, several development milestones
The complete 'my-latex' package was further divided/organized, with the
LaTeX/xparse/expl3(LaTeX3) commands and \let assignments separated from
the Plain TeX macro definitions in the 'my-definitions' subpackage and
placed in a new 'my-commands' subpackage.

The document/package hooks/patches/configuration commands were also
separated from 'my-definitions' and placed in a new 'my-config'
subpackage.   The configuration commands contained in the
document/preamble/page/auxwrite hooks/patches corresponding to the front
and main document matter were encapsulated in the new commands
'frontmatter' and 'mainmatter', respectively, and then added to the
'thedocument' key environment directly.  The remaining configurations
were then tested and reorganized according to the required insertion
point and precedence, resulting in nearly all such commands being placed
in the preamble hooks with the few remaining commands were placed in the
document environment hooks. This removed the need for any configurations
executed with respect to the beginning/end of the 'coverpage'
environment so these are no longer needed.

Separate configuration settings/commands were also created for the Table
of Contents content, such as the linespread settings and geometry,
ensuring that the coverpage and ToC apply the appropriate settings IF
they are included in the document and exclusion of one or both of them
does not affect/interfere with the settings/configurations of any other
document content.

All commands/routines/schemes previously in development were also
completed and the corresponding control sequences moved to
'my-commands'.  This includes
(1) the extension of the # to text conversion to the creation of control
sequences which are named based on the text equivalent of a provided
number,
(2) extension and/or combination of the argument list parsing and
\setlength,\let,\def operation commands such that the same operation can
still be applied to each argument or argument pair with the same sytax
but the same command can now also be used to perform mized operations
using a 2nd (3rd for pairs) list indifcating the operation to perform
for each argument (pair).
(3) Various types of lists can now be parsed N arguments at a time and
elements of a list can now be accessed by index, including cyclically
using the modular arithmetic reduction of the index value.
(4) New simplified commands were created for conditionals involving
non-text indicator types since these can be differentiated
automatically, with the common portions of the new and existing commands
replaced by newly created helper commands and used to create additional
new commands to expand the provided true or false statement (\ifTFdo)
and set their value based on an explicit value or the value of a 2nd
indicator (\setTFvar) based on the provided indicator's value.  These
indicators can also be used as the control for the \do@once command
which combines these 2 commands to ensure the conditional statement is
only executed once per run.  Those commands performing the equivalent
operations as the existing commands are simply wrappers for these
commands, determining the target indicator type and calling the existing
commands with the corresponding arguments provided automatically.  To
avoid usage of different commands dependong on indicator type, the new
non-text indicator commands will be modified to allow additional
arguments and pass these and text-based indicators directly through the
the previousy existing commands that support all indicators.  Eventually
it will also be possible to control T/F statement expansion based on
text indicators as well.  Since text indicators are primarily intended
as text output and not conditional control, lack of this capability is
not an immediate concern.
(5) Additional less important developments are underlining commands that
create a gap around letters that pass through the underline, creation of
a list from a single item via duplication,encapsulation of the underline
settings to prevent nested key command clashes, new commands to simplify
the creation of commands with a starred version, automatically choosing
between \newcommand and \renewcommand, providing the result of an
operation with C syntax return=X;, and definition of \global versions of
the macros used in the conditionals and register variable manipulation
commands whose operations depend on indicator type and register
operation, respectively, since these do not behave as desired in cases
where they are used in a group or local scope.
(6) Lastly, a large amount of legacy code no longer used has been
removed and signifiicant progress has been made towards organizing the
subpackages.