Massive Time Interval and Development Commit
This is the first commit in 8 months, there are too many changes to
mention.  A second commit will follow shortly, this one corresponds to
the current state of the package directly from the GitHub repository.
Recent development (last 1-2 weeks) has been to add support for querying
system and TeX related information at the time and location where the
package is being used.  This includes operating system, system and TeX
environement variables and package/file/system paths, the TeX
distribution/text editor/compiler, etc..  The purpose is to ease
document collaborations, particularly those shared through dropbox, and
providing the ability to use this package without requiring it to be
installed in each person's TeX library.  Many collaborators are not
comfortable with LaTeX itself and would find it difficult to add the
package to their system, so there are now new package options for
specifying usage of the package from the library or direct interaction
with its files in a local directory, with a number of various methods
for accomplishing this.

There have also been many functional improvements of the package,
particularly in terms of creating a base set of a surprisingly small
number of generic macros from which nearly all commands/environements
can be constructed.  This package provides commands/environments for
each type of possible content, which corresponds to
figures/tables/diagrams, code listings, enumerated lists,
problems/solutions, defintitions/theorems/etc., document
chapters/sections/etc., standard text/paragraphs, and front/back matter
(Table of Contents/Index/Glossary/etc.).  Each type of document content
has a single or small # of corresponding commands/environments with an
additional layer of layout/appearance controls/options.  The
layout/appearance includes placing frames around content or placing the
content within boxes and control of the frame/background/text color and
style, with each type of content having its own default design.  Each
content type has an interface of primarily key=value options to choose
the overall design scheme from a set of predefined configurations and
then change even the smallest details but also providing the ability to
build a new scheme completely from scratch.

The package had grown unwieldy given the complexity of the content
commands/environments particularly given the level of control over
detail given to the user, which is why the level of abstraction
performed to make it possible to build these from the same common set of
tasks, thereby greatly simplifying and reducing the volume of code and
making it much easier to identify and resolve issues.

The abstraction process needs to be performed again in the reverse
direction to really make this package ready for TeX distribution.  The
commands/environments were broken into smaller pieces and as common
tasks were identified, these were written into control sequence such
that everyone command/environment that performed this task could use it.
This became an iterative process of breaking tasks into smaller basic
tasks and occasionally rewriting/redesigning these as a different better
modulization was identified that would apply more broadly or be more
flexible/robust/safe.  Once the various system/TeX interaction commands
have been completed, it is now time to analyze the smallest task
commands and determine how best to construct the desired
commands/environments from the bottom up rather than starting with the
composite commands/environments with a potentially flawed design and
determining what I can find in common with others.

This commit only has part of the new system/TeX interaction commands
because the recent changes were made to the copies of the package files
placed in the .tex document's local directory and this development
begain with copying the files from the repository into this local
directory and then changing these files directly since they are the ones
being loaded.  The package files in the TeX distribution library are
links to the corresponding files in the repository so modifications to
the repository are immediately available to the library and version
control can be maintained from a single location.  Maintaining multiple
separate copies of files is always problematic and has been a serious
issue during the current development.  Having been focused on the file
and system/shell IO interactions I had not noticed that changes had
caused issued and broken several aspects of the typeset content and
attempting to find the copy of each file before/after the issues
presented themselves and backtracking without losing the advancements
made on other aspects was difficult and time consuming.  Regular commits
had not been followed because the GitHub IDE and shell applications both
stopped working and this incident forced this to be resolved after
months of delay, so having now fixed that (firewall issue) and the
content issues these files can now finallly be committed.  Ultimately
the goal is to be able to have local package file usage on Dropbox etc
done via repository links as well, though this will only be possible if
Dropbox's synchronization process follows these links such that the
current status of the repository file appears in the Dropbox folder of
those its being shared with.  The files being committed here only have
the initial steps in the local package usage development and the
subsequent commit will have those of the current status, but this too
needs additional work and testing the ability to link repo and Dropbox
files can only be really tested upon finishing this phase of
development.  This linking presents an entirely new level of
complications given that many of the system/TeX interactions involve
relative paths and it is unclear which file locations such path
information will be relative to, so links may need to be reversed but
this will be investigated when the underlying procedures have been
developed and thoroughly tested first.

and drammatically improving the debugging process.