Basic local package usage framework
This commit has the foundation of procedures for using this package by
placing its files in the directory of the .tex file being created with
it and explicitly loading these files rather than attempting to load the
package from the library.  This process includes many system/TeX
interactions accomplished via file IO and shell command executions to
determine environment variable values and absolute/relative paths to
various files.  Much of this varies depending on Windows/Unix OS, not
just in values, but procedures vary as well, so this must be determined
before such tasks can be performed.  The current status of development
is a package can now be used without it needing to be installed in the
library and all components have been thoroughly tested, the procedure is
now being cleaned up and expanded on to accumulate and organize the
gathered input information before moving on to determine how and how
much the procedure must vary to accomodate different TeX
Editor/Distribution and computer system.  It has been designed to adjust
to such differences but this needs to be investigated.  This commit
therefore corresponds to the functional version of the procedure after
having just finished fixing the issues arising in the process and the
next commit(s) will track the refinement of this approach, ideally as a
aingle commit if no significant issues arise, at which point development
will proceed to tests with different computer systems and
configurations.