Minor library package update
In developing the ability to use this package without requiring it to be
installed by placing it in a subdirectory of the .tex document which
will use it, nearly all modifications of the package files have been
made to their copies placed in the dropbox folder of the "Reminder
Walker" project.  Once this local package capability was made to work,
additional changes were then made to these local files to simplify and
clean up the implementation that made this work.  This resulted in the
insertion of some errors that made the standard library package option
no longer functional and while debugging this there were some changes
made to the lirbary package files as well.

These errors have been corrected and both the library and local
subdirectory package options are working (and presumably the user
specified directory location option).  The next step is to now commit
the changes made to the package files in the local dropbox subdirectory
of the project so the library package files will be updated to include
this additional functionality, but to make sure the relatveily minor
changes to the library package files are not lost in case errors arise
in this process, the current state of the library package files are
being committed so it is possible to return to the current working state
of all the package files.